did some wacky stuff
What this *should* do is create a command that allows me to set the specific number of leds to be red, however it might not work exactly, but it should work. This is mainly just cuz we use the light subsystem a little differently than the others cuz state machines are pog.
-Spencer